requirements.txt
asyncpg
ollama
fastapi 
uvicorn
python-dotenv>=1.0.1, <2.0.0

openai>=0.28.0,<0.29.0
pandas>=2.1.0,<3.0.0
matplotlib>=3.7.2,<4.0.0
plotly>=5.16.1,<5.17.0
scipy>=1.11.2,<2.0.0
scikit-learn>=1.3.0,<2.0.0
tiktoken>=0.4.0,<0.5.0
google-api-python-client>=2.98.0,<3.0.0
youtube-transcript-api>=0.6.1,<1.0.0
rich>=13.5.2,<14.0.0
tenacity>=8.2.3